+++ src/lib/mympd_queue.h
@@ -31,5 +31,5 @@ void mympd_queue_free(struct t_mympd_queue *queue);
 int mympd_queue_push(struct t_mympd_queue *queue, void *data, long id);
 void *mympd_queue_shift(struct t_mympd_queue *queue, int timeout, long id);
 void *mympd_queue_expire(struct t_mympd_queue *queue, time_t max_age);
-unsigned mympd_queue_length(struct t_mympd_queue *queue, int timeout);
+long mympd_queue_length(struct t_mympd_queue *queue, int timeout);
 #endif
